LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.isPointInPath.basic.1-expected.txt

2d.path.isPointInPath.basic.1

isPointInPath() detects whether the point is inside the path


PASS isPointInPath() detects whether the point is inside the path
